Many tweaks
Implemented a better serialization system that allows for more control
and probably higher performance. However, the client can no longer
corrently load the world and run the game, so I'll be looking into a
fix. A new class ServerGameUpdater will replace much of the server code
to make the server require less synchronization which should lead to
less bugginess. It uses a message passing system that allows the server
to update the game state when it receives messages from clients. The
messages, called GameTasks, are processing each game update loop.